requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
numpy==2.1.1
Pillow==10.4.0